requirements.txt
fire
openai
prettytable
numpy
sklearn
google-generativeai
tqdm